+++ component/src/test/java/arez/component/MemoizeCacheTest.java
@@ -11,6 +11,7 @@ import arez.spy.ComponentInfo;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.realityforge.guiceyloops.shared.ValueUtil;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
@@ -18,6 +19,16 @@ import static org.testng.Assert.*;
 public class MemoizeCacheTest
   extends AbstractArezComponentTest
 {
+  @BeforeMethod
+  @Override
+  protected void beforeTest()
+    throws Exception
+  {
+    super.beforeTest();
+    ArezTestUtil.disableZones();
+    ArezTestUtil.resetState();
+  }
+
   @Test
   public void basicOperation()
   {
@@ -30,7 +41,7 @@ public class MemoizeCacheTest
     final ArezContext context = Arez.context();
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( context, null, name, function, 2 );
+      new MemoizeCache<>( null, null, name, function, 2 );
 
     assertEquals( cache.isDisposed(), false );
     assertEquals( cache.getNextIndex(), 0 );
@@ -101,10 +112,9 @@ public class MemoizeCacheTest
       callCount.incrementAndGet();
       return args[ 0 ] + "." + args[ 1 ];
     };
-    final ArezContext context = Arez.context();
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( context, null, name, function, 2 );
+      new MemoizeCache<>( null, null, name, function, 2 );
 
     final ComputedValue<String> computedValue1 = cache.getComputedValue( "a", "b" );
     final ComputedValue<String> computedValue1b = cache.getComputedValue( "a", "b" );
@@ -138,7 +148,7 @@ public class MemoizeCacheTest
 
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( context, component, name, function, 2 );
+      new MemoizeCache<>( null, component, name, function, 2 );
 
     final ComputedValue<String> computedValue1 = cache.getComputedValue( "a", "b" );
 
@@ -165,7 +175,7 @@ public class MemoizeCacheTest
     final ArezContext context = Arez.context();
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( context, null, name, function, 2 );
+      new MemoizeCache<>( null, null, name, function, 2 );
 
     assertEquals( cache.isDisposed(), false );
     assertEquals( cache.getNextIndex(), 0 );
@@ -207,7 +217,7 @@ public class MemoizeCacheTest
     final ArezContext context = Arez.context();
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( context, null, name, function, 2 );
+      new MemoizeCache<>( null, null, name, function, 2 );
 
     assertEquals( cache.getNextIndex(), 0 );
 
@@ -224,7 +234,7 @@ public class MemoizeCacheTest
   public void disposeComputedValue_passedBadArgCounts()
   {
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( Arez.context(), null, ValueUtil.randomString(), args -> args[ 0 ] + "." + args[ 1 ], 2 );
+      new MemoizeCache<>( null, null, ValueUtil.randomString(), args -> args[ 0 ] + "." + args[ 1 ], 2 );
 
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class, () -> cache.disposeComputedValue( "a" ) );
@@ -237,7 +247,7 @@ public class MemoizeCacheTest
   public void get_passedBadArgCounts()
   {
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( Arez.context(), null, ValueUtil.randomString(), args -> args[ 0 ] + "." + args[ 1 ], 2 );
+      new MemoizeCache<>( null, null, ValueUtil.randomString(), args -> args[ 0 ] + "." + args[ 1 ], 2 );
 
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class, () -> cache.get( "a" ) );
@@ -250,7 +260,7 @@ public class MemoizeCacheTest
   public void get_invokedOnDisposed()
   {
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( Arez.context(), null, "X", args -> args[ 0 ] + "." + args[ 1 ], 2 );
+      new MemoizeCache<>( null, null, "X", args -> args[ 0 ] + "." + args[ 1 ], 2 );
 
     cache.dispose();
 
@@ -267,7 +277,7 @@ public class MemoizeCacheTest
 
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class,
-                    () -> new MemoizeCache<>( Arez.context(), null, "X", args -> args[ 0 ], 1 ) );
+                    () -> new MemoizeCache<>( null, null, "X", args -> args[ 0 ], 1 ) );
 
     assertEquals( exception.getMessage(),
                   "Arez-0159: MemoizeCache passed a name 'X' but Arez.areNamesEnabled() is false" );
@@ -278,9 +288,22 @@ public class MemoizeCacheTest
   {
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class,
-                    () -> new MemoizeCache<>( Arez.context(), null, "X", args -> args[ 0 ], 0 ) );
+                    () -> new MemoizeCache<>( null, null, "X", args -> args[ 0 ], 0 ) );
 
     assertEquals( exception.getMessage(),
                   "Arez-0160: MemoizeCache constructed with invalid argCount: 0. Expected positive value." );
+  }
+
+  @Test
+  public void constructorPassedContext_whenZonesDisabled()
+  {
+    ArezTestUtil.disableZones();
+    ArezTestUtil.resetState();
+    final IllegalStateException exception =
+      expectThrows( IllegalStateException.class,
+                    () -> new MemoizeCache<>( Arez.context(), null, null, args -> args[ 0 ], 1 ) );
+
+    assertEquals( exception.getMessage(),
+                  "Arez-0168: MemoizeCache passed a context but Arez.areZonesEnabled() is false" );
   }
 }